Add basic feed endpoint for #820
