+++ Core/Providers/APIProvider.php
@@ -7,6 +7,7 @@ use Intraxia\Jaxion\Assets\Register as Assets;
 use Intraxia\Jaxion\Assets\ServiceProvider as ServiceProvider;
 
 use PressForward\Core\API\PostExtension;
+use PressForward\Core\API\FeedEndpoint;
 
 
 class APIProvider extends ServiceProvider {
@@ -16,6 +17,12 @@ class APIProvider extends ServiceProvider {
 			'api.post_extension',
 			function( $container ){
 				return new PostExtension( $container->fetch('controller.metas') );
+			}
+		);
+		$container->share(
+			'api.feed_endpoint',
+			function( $container ){
+				return new FeedEndpoint( $container->fetch('controller.metas') );
 			}
 		);
 	}
